+++ src/components/Report/Report.module.scss
@@ -3,8 +3,12 @@
 }
 .balance_segment {
   position: absolute;
-  top: 0;
+  margin-left: auto;
+  margin-right: auto;
+  top: 15px;
   left: 0;
+  right: 0;
+  text-align: center;
 }
 
 .balance_line {
@@ -14,8 +18,11 @@
   justify-content: center;
   padding-top: 20px;
   padding-bottom: 20px;
-  width: 100%;
   margin-left: 0%;
+}
+
+.return_home_div {
+  margin-left: 10px;
 }
 
 @media screen and (max-width: 767px) {
@@ -24,11 +31,15 @@
     flex-direction: column;
     justify-content: center; */
   }
-  .placeholder,
-  .return_home_div,
+  .return_home_div {
+    margin-bottom: 10px;
+  }
+
   .month_selector_div {
-    padding-top: 10px;
-    padding-bottom: 10px;
+    margin-bottom: 30px;
+  }
+  .placeholder {
+    margin-bottom: 30px;
   }
   .return_home_div {
     padding-left: 10px;
@@ -39,30 +50,42 @@
     justify-content: center;
   }
   .balance_line {
-    width: 100%;
-    margin-left: 0%;
+    width: 88%;
+    margin-left: 5.9%;
+    margin-right: 5.9%;
+    margin-bottom: 40px;
+  }
+
+  .balance_component {
+    margin-bottom: 30px;
   }
 }
 
 @media screen and (min-width: 768px) {
   .control_line {
     display: flex;
-    justify-content: space-around;
+    justify-content: space-between;
+    align-items: center;
     padding-top: 20px;
-    padding-bottom: 20px;
-    width: 80%;
-    margin-left: 10%;
+    margin-right: 6.5%;
+    margin-left: 6.5%;
+    margin-bottom: 40px;
   }
 
   .balance_line {
-    width: 80%;
-    margin-left: 10%;
+    margin-right: 6.5%;
+    margin-left: 6.5%;
+    margin-bottom: 40px;
   }
 }
 
 @media screen and (min-width: 1280px) {
+  .control_line {
+    margin-right: 8.6%;
+    margin-left: 8.6%;
+  }
   .balance_line {
-    width: 80%;
-    margin-left: 10%;
+    margin-right: 8.6%;
+    margin-left: 8.6%;
   }
 }
