+++ src/components/ReturnHome/ReturnHome.module.css
@@ -1,5 +1,5 @@
 .container {
-  //width: 230px;
+  /* width: 230px; */
   display: flex;
 }
